requirements.txt
fastapi
uvicorn
pymysql
sqlalchemy
boto3
python-multipart
pymysql
pydantic[email]
passlib[bcrypt]
python-dotenv
cryptography